refactor: replace getCategoryName function to helpers
